fix: repology project filtering algorithm

+++ server/listener/health_check.go
@@ -20,7 +20,7 @@ func OnServerOnline(handle func()) {
 	onServerOnlineHandlers = append(onServerOnlineHandlers, handle)
 }
 
-func triggerServerOnline(port int) {
+func triggerServerOnline(port uint16) {
 	retryDelay := 100
 	timeout := 5000
 	for ; timeout > 0; timeout -= retryDelay {
